Fixed popup issue and restructured the code
