+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "rust-analyzer.check.command": "clippy",
+    "editor.formatOnSave": true,
+
+    "[html]": {
+        "editor.defaultFormatter": "vscode.html-language-features"
+    }
+
+}
