+++ src/privacy.rs
@@ -1,0 +1,7 @@
+use crate::templates::privacy_template::PrivacyTemplate;
+use actix_web::Error;
+use actix_web::Responder;
+
+pub async fn privacy() -> Result<impl Responder, Error> {
+    Ok(PrivacyTemplate {})
+}
