+++ src/templates/mod.rs
@@ -1,0 +1,2 @@
+pub mod privacy_template;
+pub mod upload_template;
