+++ src/templates/privacy_template.rs
@@ -1,0 +1,6 @@
+use askama::Template;
+use askama_web::WebTemplate;
+
+#[derive(Template, WebTemplate)]
+#[template(path = "privacy.html")]
+pub struct PrivacyTemplate;
