+++ static/html/index.html
@@ -1,32 +1,32 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <title>Example</title>
-    <script src="/js/htmx.min.js"></script>
-    <link href="/html/output.css" rel="stylesheet" />
-  </head>
 
-  <body class="h-screen flex items-center justify-end">
-    <!--
+<head>
+  <title>Example</title>
+  <script src="/js/htmx.min.js"></script>
+  <link href="/html/output.css" rel="stylesheet" />
+</head>
+
+<body>
+  <!--
     <p class="text-amber-600 font-bold text-3xl">
       This is an example of a simple HTML page with one paragraph.
     </p>
 
     <h1 class="text-3xl font-bold underline">Hello world!</h1>-->
 
-    <!--Most modern smartphones are size md so only sm wont work-->
+  <!--Most modern smartphones are size md so only sm wont work-->
 
-    <div id="error_message"></div>
+  <div id="error_message"></div>
 
-    <form
-      hx-post="/load"
-      enctype="multipart/form-data"
-      hx-target="#error_message"
-      class="bg-gray-50 p-8 
+  <div id="modal"></div>
+
+
+  <div class="h-screen flex items-center justify-end">
+    <form hx-post="/load" enctype="multipart/form-data" hx-target="#error_message" class="bg-gray-50 p-8 
         rounded-4xl 
         shadow-lg flex flex-col gap-4 border border-gray-200
-        w-full md:w-full h-screen lg:w-1/3 sm:w-full"
-    >
+        w-full md:w-full h-screen lg:w-1/3 sm:w-full">
 
       <h1 class="text-5xl text-center font-bold"></h1>
 
@@ -36,42 +36,28 @@
         Upload limit 10MiB
       </h1>
 
-      <input
-        type="file"
-        multiple
-        name="file"
-        accept="application/pdf"
-        class="block w-full text-sm text-gray-900 border border-gray-300 rounded-lg cursor-pointer bg-gray-50 focus:outline-none focus:ring-2 focus:ring-blue-500"
-      />
+      <input type="file" multiple name="file" accept="application/pdf"
+        class="block w-full text-sm text-gray-900 border border-gray-300 rounded-lg cursor-pointer bg-gray-50 focus:outline-none focus:ring-2 focus:ring-blue-500" />
 
-      <input
-        type="text"
-        name="email"
-        placeholder="max@mustermann.de"
-        class="w-full rounded-lg hover:bg-blue-50 text-black focus:outline-none focus:ring-2 focus:ring-blue-500"
-      />
+      <input type="text" name="email" placeholder="max@mustermann.de"
+        class="w-full rounded-lg hover:bg-blue-50 text-black focus:outline-none focus:ring-2 focus:ring-blue-500" />
 
       <label class="flex items-center gap-2 text-gray-700">
         <input type="checkbox" name="checkbox" class="size-4 accent-blue-600" />
         <span>
-          I agree to the 
-          <a
-            href="/privacy"
-            hx-get="/privacy"
-            hx-target="#modal"
-            hx-swap="innerHTML"
-            class="text-blue-600 underline cursor-pointer"
-          >data protection agrement</a>
+          I agree to the
+          <a href="/privacy" hx-get="/privacy" hx-target="#modal" hx-swap="innerHTML"
+            class="text-blue-600 underline cursor-pointer">data protection agrement</a>
         </span>
       </label>
 
-      <button
-        type="submit"
-        class="mt-2 bg-blue-600 hover:bg-blue-700 text-white font-semibold py-2 rounded-lg transition"
-      >
+      <button type="submit"
+        class="mt-2 bg-blue-600 hover:bg-blue-700 text-white font-semibold py-2 rounded-lg transition">
         Submit
       </button>
     </form>
-  </body>
-    <div id="modal"></div>
+  </div>
+
+</body>
+
 </html>
