+++ static/html/input.css
@@ -1,1 +1,3 @@
 @import "tailwindcss";
+
+@config "../../tailwind.config.js"
