+++ tailwind.config.js
@@ -1,0 +1,6 @@
+// tailwind.config.js
+module.exports = {
+  content: ["./templates/*.html", "./static/html/index.html"], // include all HTMX-injected templates
+  theme: { extend: {} },
+  plugins: [],
+}
